+++ great_expectations/rule_based_profiler/parameter_builder/__init__.py
@@ -1,11 +1,12 @@
 from great_expectations.rule_based_profiler.parameter_builder.parameter_builder import (  # isort:skip
     AttributedResolvedMetrics,
     ParameterBuilder,
-    MetricComputationResult,
     MetricValues,
     MetricComputationDetails,
     init_rule_parameter_builders,
 )
+from great_expectations.rule_based_profiler.types.metric import MetricComputationResult
+
 from great_expectations.rule_based_profiler.parameter_builder.metric_multi_batch_parameter_builder import (  # isort:skip
     MetricMultiBatchParameterBuilder,
 )
