+++ great_expectations/rule_based_profiler/parameter_builder/parameter_builder.py
@@ -34,77 +34,18 @@ from great_expectations.rule_based_profiler.types import (
     build_parameter_container,
     get_fully_qualified_parameter_names,
 )
-from great_expectations.types import SerializableDictDot
+from great_expectations.rule_based_profiler.types.metric import (
+    AttributedResolvedMetrics,
+    MetricComputationDetails,
+    MetricComputationResult,
+    MetricValue,
+    MetricValues,
+)
 from great_expectations.types.attributes import Attributes
 from great_expectations.validator.metric_configuration import MetricConfiguration
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
-
-# TODO: <Alex>These are placeholder types, until a formal metric computation state class is made available.</Alex>
-MetricValue = Union[Any, List[Any], pd.DataFrame, pd.Series, np.ndarray]
-MetricValues = Union[MetricValue, pd.DataFrame, pd.Series, np.ndarray]
-MetricComputationDetails = Dict[str, Any]
-MetricComputationResult = make_dataclass(
-    "MetricComputationResult", ["attributed_resolved_metrics", "details"]
-)
-
-
-@dataclass
-class AttributedResolvedMetrics(SerializableDictDot):
-    """
-    This class facilitates computing multiple metrics as one operation.
-
-    In order to gather results pertaining to diverse MetricConfiguration directives, computed metrics are augmented
-    with uniquely identifiable attribution object so that receivers can filter them from overall resolved metrics.
-    """
-
-    metric_attributes: Optional[Attributes] = None
-    metric_values_by_batch_id: Optional[Dict[str, MetricValue]] = None
-
-    @staticmethod
-    def get_metric_values_from_attributed_metric_values(
-        attributed_metric_values: Dict[str, MetricValue]
-    ) -> MetricValues:
-        if attributed_metric_values is None:
-            return None
-
-        values: MetricValues = list(attributed_metric_values.values())[0]
-        if values is not None and isinstance(values, (pd.DataFrame, pd.Series)):
-            return list(attributed_metric_values.values())
-
-        return np.array(list(attributed_metric_values.values()))
-
-    def add_resolved_metric(self, batch_id: str, value: MetricValue) -> None:
-        if self.metric_values_by_batch_id is None:
-            self.metric_values_by_batch_id = {}
-
-        if not isinstance(self.metric_values_by_batch_id, OrderedDict):
-            self.metric_values_by_batch_id = OrderedDict(self.metric_values_by_batch_id)
-
-        self.metric_values_by_batch_id[batch_id] = value
-
-    @property
-    def id(self) -> str:
-        return self.metric_attributes.to_id()
-
-    @property
-    def attributed_metric_values(self) -> Optional[Dict[str, MetricValue]]:
-        return self.metric_values_by_batch_id
-
-    @property
-    def metric_values(self) -> MetricValues:
-        return (
-            AttributedResolvedMetrics.get_metric_values_from_attributed_metric_values(
-                attributed_metric_values=self.attributed_metric_values
-            )
-        )
-
-    def to_dict(self) -> dict:
-        return asdict(self)
-
-    def to_json_dict(self) -> dict:
-        return convert_to_json_serializable(data=self.to_dict())
 
 
 class ParameterBuilder(ABC, Builder):
@@ -528,6 +469,7 @@ specified (empty "metric_name" value detected)."""
                     batch_id: [resolved_metric_value]
                     for batch_id, resolved_metric_value in attributed_resolved_metrics.attributed_metric_values.items()
                 }
+                print("hello stop me here")
                 attributed_resolved_metrics_map[
                     metric_attributes_id
                 ] = attributed_resolved_metrics
@@ -549,7 +491,8 @@ specified (empty "metric_name" value detected)."""
             )
 
         # Step-10: Build and return result to receiver (apply simplifications to cases of single "metric_value_kwargs").
-
+        # is this where we need to do a bit of clean up?
+        # how deep do we go?
         return MetricComputationResult(
             attributed_resolved_metrics=list(attributed_resolved_metrics_map.values()),
             details={
